[17.0][FIX] website_sale_hide_price: qweb cache
Fix caching of the website_sale.products view override.

Before this commit, this module added a view that added two price website related cache keys to the products grid wrapper. In Odoo 17.0, caching was removed for this element. In combination with this module, this resulted in re-rendering this portion of the view, only depending on the cache keys added by this module. Because of it, searched products didn't update when changing any of the search parameters.

This commit fixes it, by removing that override, and leaving the elements cache disabled.